FIXED: Many warnings with gcc 4.3. I did not catch all yet, but I think I got enough to allow in a release when merged back to branch. I will continue to fix the rest in trunk sometime.
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@858 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ test/PdfString_Test.cpp
@@ -157,7 +157,7 @@ void testUnicode()
 void testString( const char* pszString, const PdfString & str, const PdfString & hex ) 
 {
     printf("\t->    Got string: %s\n", pszString );
-    printf("\t-> ... of length: %li\n", strlen( pszString ) );
+    printf("\t-> ... of length: %li\n", static_cast<long>(strlen( pszString )) );
     printf("\t-> Got PdfString: %s\n", str.GetString() );
     printf("\t-> ... of length: %li\n", str.GetLength() );
     printf("\t-> Got hexstring: %s\n", hex.GetString() );
@@ -171,7 +171,7 @@ void testString( const char* pszString, const PdfString & str, const PdfString &
 
     if( static_cast<size_t>(str.GetLength()) != strlen( pszString ) ) 
     {
-        printf("Strings length is not equal: %li vs %li!\n", str.GetLength(), strlen( pszString ) );
+        printf("Strings length is not equal: %li vs %li!\n", str.GetLength(), static_cast<long>(strlen( pszString )) );
         PODOFO_RAISE_ERROR( ePdfError_TestFailed );
     }
